+++ src/WatsonWebsocket/WatsonWsServer.cs
@@ -44,12 +44,12 @@ namespace WatsonWebsocket
         /// <summary>
         /// Event fired when a client connects.
         /// </summary>
-        public event EventHandler<ClientConnectedEventArgs> ClientConnected;
+        public event EventHandler<ConnectionEventArgs> ClientConnected;
 
         /// <summary>
         /// Event fired when a client disconnects.
         /// </summary>
-        public event EventHandler<ClientDisconnectedEventArgs> ClientDisconnected;
+        public event EventHandler<DisconnectionEventArgs> ClientDisconnected;
 
         /// <summary>
         /// Event fired when the server stops.
@@ -111,8 +111,8 @@ namespace WatsonWebsocket
         private List<string> _ListenerPrefixes = new List<string>();
         private HttpListener _Listener;
         private readonly object _PermittedIpsLock = new object();
-        private ConcurrentDictionary<string, ClientMetadata> _Clients; 
-        private CancellationTokenSource _TokenSource;
+        private ConcurrentDictionary<Guid, ClientMetadata> _Clients = new ConcurrentDictionary<Guid, ClientMetadata>();
+        private CancellationTokenSource _TokenSource = new CancellationTokenSource();
         private CancellationToken _Token; 
         private Task _AcceptConnectionsTask;
         private Statistics _Stats = new Statistics();
@@ -139,13 +139,9 @@ namespace WatsonWebsocket
 
             _Listener = new HttpListener();
             foreach (string prefix in _ListenerPrefixes)
-            {
                 _Listener.Prefixes.Add(prefix);
-            }
 
-            _TokenSource = new CancellationTokenSource();
             _Token = _TokenSource.Token;
-            _Clients = new ConcurrentDictionary<string, ClientMetadata>();
         }
 
         /// <summary>
@@ -173,9 +169,7 @@ namespace WatsonWebsocket
                 _Listener.Prefixes.Add(prefix);
             }
 
-            _TokenSource = new CancellationTokenSource();
             _Token = _TokenSource.Token;
-            _Clients = new ConcurrentDictionary<string, ClientMetadata>();
         }
 
         /// <summary>
@@ -194,13 +188,9 @@ namespace WatsonWebsocket
             {
                 var dnsLookup = Dns.GetHostEntry(uri.Host);
                 if (dnsLookup.AddressList.Length > 0)
-                {
                     host = dnsLookup.AddressList.First().ToString();
-                }
                 else
-                {
                     throw new ArgumentException("Cannot resolve address to IP.");
-                }
             }
             else
             {
@@ -216,13 +206,9 @@ namespace WatsonWebsocket
 
             _Listener = new HttpListener();
             foreach (string prefix in _ListenerPrefixes)
-            {
                 _Listener.Prefixes.Add(prefix);
-            }
 
-            _TokenSource = new CancellationTokenSource();
             _Token = _TokenSource.Token;
-            _Clients = new ConcurrentDictionary<string, ClientMetadata>();
         }
 
         #endregion
@@ -300,69 +286,61 @@ namespace WatsonWebsocket
         /// <summary>
         /// Send text data to the specified client, asynchronously.
         /// </summary>
-        /// <param name="ipPort">IP:port of the recipient client.</param>
+        /// <param name="guid">Globally-unique identifier of the recipient client.</param>
         /// <param name="data">String containing data.</param>
         /// <param name="token">Cancellation token allowing for termination of this request.</param>
         /// <returns>Task with Boolean indicating if the message was sent successfully.</returns>
-        public Task<bool> SendAsync(string ipPort, string data, CancellationToken token = default)
+        public Task<bool> SendAsync(Guid guid, string data, CancellationToken token = default)
         {
-            if (String.IsNullOrEmpty(data)) throw new ArgumentNullException(nameof(data));
-            if (!_Clients.TryGetValue(ipPort, out ClientMetadata client))
-            {
-                Logger?.Invoke(_Header + "unable to find client " + ipPort);
-                return Task.FromResult(false);
-            }
-
-            Task<bool> task = MessageWriteAsync(client, new ArraySegment<byte>(Encoding.UTF8.GetBytes(data)), WebSocketMessageType.Text, token);
-
-            client = null;
-            return task;
+            if (String.IsNullOrEmpty(data)) data = "";
+            return SendAsync(guid, Encoding.UTF8.GetBytes(data), token);
         }
 
         /// <summary>
         /// Send binary data to the specified client, asynchronously.
         /// </summary>
-        /// <param name="ipPort">IP:port of the recipient client.</param>
+        /// <param name="guid">Globally-unique identifier of the recipient client.</param>
         /// <param name="data">Byte array containing data.</param> 
         /// <param name="token">Cancellation token allowing for termination of this request.</param>
         /// <returns>Task with Boolean indicating if the message was sent successfully.</returns>
-        public Task<bool> SendAsync(string ipPort, byte[] data, CancellationToken token = default)
+        public Task<bool> SendAsync(Guid guid, byte[] data, CancellationToken token = default)
         {
-            return SendAsync(ipPort, new ArraySegment<byte>(data), WebSocketMessageType.Binary, token);
+            if (data == null) data = new byte[0];
+            return SendAsync(guid, new ArraySegment<byte>(data), WebSocketMessageType.Binary, token);
         }
 
         /// <summary>
         /// Send binary data to the specified client, asynchronously.
         /// </summary>
-        /// <param name="ipPort">IP:port of the recipient client.</param>
+        /// <param name="guid">Globally-unique identifier of the recipient client.</param>
         /// <param name="data">Byte array containing data.</param> 
         /// <param name="msgType">Web socket message type.</param>
         /// <param name="token">Cancellation token allowing for termination of this request.</param>
         /// <returns>Task with Boolean indicating if the message was sent successfully.</returns>
-        public Task<bool> SendAsync(string ipPort, byte[] data, WebSocketMessageType msgType, CancellationToken token = default)
+        public Task<bool> SendAsync(Guid guid, byte[] data, WebSocketMessageType msgType, CancellationToken token = default)
         {
-            return SendAsync(ipPort, new ArraySegment<byte>(data), msgType, token);
+            if (data == null) data = new byte[0];
+            return SendAsync(guid, new ArraySegment<byte>(data), msgType, token);
         }
 
         /// <summary>
         /// Send binary data to the specified client, asynchronously.
         /// </summary>
-        /// <param name="ipPort">IP:port of the recipient client.</param>
+        /// <param name="guid">Globally-unique identifier of the recipient client.</param>
         /// <param name="data">ArraySegment containing data.</param> 
         /// <param name="msgType">Web socket message type.</param>
         /// <param name="token">Cancellation token allowing for termination of this request.</param>
         /// <returns>Task with Boolean indicating if the message was sent successfully.</returns>
-        public Task<bool> SendAsync(string ipPort, ArraySegment<byte> data, WebSocketMessageType msgType = WebSocketMessageType.Binary, CancellationToken token = default)
+        public Task<bool> SendAsync(Guid guid, ArraySegment<byte> data, WebSocketMessageType msgType = WebSocketMessageType.Binary, CancellationToken token = default)
         {
             if (data.Array == null || data.Count < 1) throw new ArgumentNullException(nameof(data));
-            if (!_Clients.TryGetValue(ipPort, out ClientMetadata client))
+            if (!_Clients.TryGetValue(guid, out ClientMetadata client))
             {
-                Logger?.Invoke(_Header + "unable to find client " + ipPort);
+                Logger?.Invoke(_Header + "unable to find client " + guid.ToString());
                 return Task.FromResult(false);
             }
 
             Task<bool> task = MessageWriteAsync(client, data, msgType, token);
-
             client = null;
             return task;
         }
@@ -370,29 +348,29 @@ namespace WatsonWebsocket
         /// <summary>
         /// Determine whether or not the specified client is connected to the server.
         /// </summary>
-        /// <param name="ipPort">IP:port of the recipient client.</param>
+        /// <param name="guid">Globally-unique identifier of the recipient client.</param>
         /// <returns>Boolean indicating if the client is connected to the server.</returns>
-        public bool IsClientConnected(string ipPort)
+        public bool IsClientConnected(Guid guid)
         {
-            return _Clients.TryGetValue(ipPort, out _);
+            return _Clients.Any(c => c.Key.Equals(guid));
         }
 
         /// <summary>
-        /// List the IP:port of each connected client.
+        /// Retrieve list of connected clients.
         /// </summary>
-        /// <returns>A string list containing each client IP:port.</returns>
-        public IEnumerable<string> ListClients()
+        /// <returns>A list containing client metadata.</returns>
+        public IEnumerable<ClientMetadata> ListClients()
         {
-            return _Clients.Keys.ToArray();
+            return _Clients.Values;
         }
 
         /// <summary>
         /// Forcefully disconnect a client.
         /// </summary>
-        /// <param name="ipPort">IP:port of the client.</param>
-        public void DisconnectClient(string ipPort)
+        /// <param name="guid">Globally-unique identifier of the recipient client.</param>
+        public void DisconnectClient(Guid guid)
         { 
-            if (_Clients.TryGetValue(ipPort, out var client))
+            if (_Clients.TryGetValue(guid, out var client))
             {
                 lock (client)
                 {
@@ -427,7 +405,7 @@ namespace WatsonWebsocket
             {
                 if (_Clients != null)
                 {
-                    foreach (KeyValuePair<string, ClientMetadata> client in _Clients)
+                    foreach (KeyValuePair<Guid, ClientMetadata> client in _Clients)
                     {
                         client.Value.Ws.CloseAsync(WebSocketCloseStatus.NormalClosure, "", client.Value.TokenSource.Token);
                         client.Value.TokenSource.Cancel();
@@ -531,9 +509,9 @@ namespace WatsonWebsocket
                             WebSocket ws = wsContext.WebSocket;
                             ClientMetadata md = new ClientMetadata(ctx, ws, wsContext, tokenSource);
                              
-                            _Clients.TryAdd(md.IpPort, md);
+                            _Clients.TryAdd(md.Guid, md);
 
-                            ClientConnected?.Invoke(this, new ClientConnectedEventArgs(md.IpPort, ctx.Request));
+                            ClientConnected?.Invoke(this, new ConnectionEventArgs(md, ctx.Request));
                             await Task.Run(() => DataReceiver(md), token);
                              
                         }, token);
@@ -559,6 +537,10 @@ namespace WatsonWebsocket
             {
                 // thrown when disposed
             }
+            catch (HttpListenerException)
+            {
+                // thrown when stopped
+            }
             catch (Exception e)
             {
                 Logger?.Invoke(_Header + "listener exception:" + Environment.NewLine + e.ToString());
@@ -569,9 +551,9 @@ namespace WatsonWebsocket
             }
         }
 
-        private async Task DataReceiver(ClientMetadata md)
+        private async Task DataReceiver(ClientMetadata client)
         { 
-            string header = "[WatsonWsServer " + md.IpPort + "] ";
+            string header = "[WatsonWsServer " + client.Guid.ToString() + "] ";
             Logger?.Invoke(header + "starting data receiver");
             byte[] buffer = new byte[65536];
 
@@ -579,7 +561,7 @@ namespace WatsonWebsocket
             { 
                 while (true)
                 {
-                    MessageReceivedEventArgs msg = await MessageReadAsync(md, buffer).ConfigureAwait(false);
+                    MessageReceivedEventArgs msg = await MessageReadAsync(client, buffer).ConfigureAwait(false);
 
                     if (msg != null)
                     {
@@ -591,7 +573,7 @@ namespace WatsonWebsocket
 
                         if (msg.Data != null)
                         {
-                            Task unawaited = Task.Run(() => MessageReceived?.Invoke(this, msg), md.TokenSource.Token);
+                            Task unawaited = Task.Run(() => MessageReceived?.Invoke(this, msg), client.TokenSource.Token);
                         }
                         else
                         {
@@ -618,17 +600,17 @@ namespace WatsonWebsocket
             }
             finally
             { 
-                string ipPort = md.IpPort;
-                ClientDisconnected?.Invoke(this, new ClientDisconnectedEventArgs(md.IpPort));
-                md.Ws.Dispose();
+                string ipPort = client.IpPort;
+                ClientDisconnected?.Invoke(this, new DisconnectionEventArgs(client));
+                client.Ws.Dispose();
                 Logger?.Invoke(header + "disconnected");
-                _Clients.TryRemove(ipPort, out _);
+                _Clients.TryRemove(client.Guid, out _);
             }
         }
          
-        private async Task<MessageReceivedEventArgs> MessageReadAsync(ClientMetadata md, byte[] buffer)
+        private async Task<MessageReceivedEventArgs> MessageReadAsync(ClientMetadata client, byte[] buffer)
         {
-            string header = "[WatsonWsServer " + md.IpPort + "] ";
+            string header = "[WatsonWsServer " + client.Guid.ToString() + "] ";
 
             using (MemoryStream ms = new MemoryStream())
             {
@@ -636,21 +618,21 @@ namespace WatsonWebsocket
 
                 while (true)
                 {
-                    WebSocketReceiveResult result = await md.Ws.ReceiveAsync(seg, md.TokenSource.Token).ConfigureAwait(false);
+                    WebSocketReceiveResult result = await client.Ws.ReceiveAsync(seg, client.TokenSource.Token).ConfigureAwait(false);
                     if (result.CloseStatus != null)
                     {
                         Logger?.Invoke(header + "close received");
-                        await md.Ws.CloseAsync(WebSocketCloseStatus.NormalClosure, "", CancellationToken.None);
+                        await client.Ws.CloseAsync(WebSocketCloseStatus.NormalClosure, "", CancellationToken.None);
                         throw new WebSocketException("Websocket closed.");
                     }
 
-                    if (md.Ws.State != WebSocketState.Open)
+                    if (client.Ws.State != WebSocketState.Open)
                     {
                         Logger?.Invoke(header + "websocket no longer open");
                         throw new WebSocketException("Websocket closed.");
                     }
 
-                    if (md.TokenSource.Token.IsCancellationRequested)
+                    if (client.TokenSource.Token.IsCancellationRequested)
                     {
                         Logger?.Invoke(header + "cancel requested");
                     }
@@ -662,7 +644,7 @@ namespace WatsonWebsocket
 
                     if (result.EndOfMessage)
                     {
-                        return new MessageReceivedEventArgs(md.IpPort, new ArraySegment<byte>(ms.GetBuffer(), 0, (int)ms.Length), result.MessageType);
+                        return new MessageReceivedEventArgs(client, new ArraySegment<byte>(ms.GetBuffer(), 0, (int)ms.Length), result.MessageType);
                     }
                 }
             }